# docs/api-spec.yaml
openapi: 3.0.0
paths:
...
  /store:
    get:
      operationId: StoreController_get
      parameters:
        - name: raw
          required: false
          in: query
          schema:
            type: string
        - name: lang
          required: false
          in: query
          schema: {}
      responses:
        '200':
          description: ''
      tags:
        - store
    patch:
      operationId: StoreController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStoreDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - store
...
  /wishlist:
    get:
      operationId: WishlistController_list
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - wishlist
  /wishlist/items:
    post:
      operationId: WishlistController_addItem
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWishlistItemDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - wishlist
  /wishlist/items/{productId}:
    delete:
      operationId: WishlistController_removeItem
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - wishlist
  /orders:
    post:
      operationId: OrdersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: ''
      security:
        - bearer: []
      tags:
        - orders
    get:
      operationId: OrdersController_list
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - orders
  /orders/{id}:
    get:
      operationId: OrdersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - orders
  /orders/{id}/cancel:
    patch:
      operationId: OrdersController_cancel
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - orders
  /orders/{id}/status:
    patch:
      operationId: OrdersController_updateStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
      responses:
        '200':
          description: ''
      security:
        - bearer: []
      tags:
        - orders
...
info:
  title: MarketApp API
  description: MarketApp backend API dokümantasyonu
  version: 0.1.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateStoreDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 150
        city:
          type: string
          maxLength: 100
        tagline:
          type: object
          description: Çok dilli slogan
        description:
          type: object
          description: Çok dilli açıklama
        logo_url:
          type: string
        cover_image_url:
          type: string
        phone:
          type: string
          maxLength: 20
        address:
          type: string
        working_hours:
          type: object
          description: Gün bazlı çalışma saatleri (serbest JSON)
...
    AddWishlistItemDto:
      type: object
      properties:
        product_id:
          type: string
      required:
        - product_id
    OrderItemInputDto:
      type: object
      properties:
        product_id:
          type: string
        quantity:
          type: number
      required:
        - product_id
        - quantity
    CreateOrderDto:
      type: object
      properties:
        address_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemInputDto'
        coupon_code:
          type: string
      required:
        - address_id
        - items
    CancelOrderDto:
      type: object
      properties:
        reason:
          type: string
          description: İptal nedeni (opsiyonel)
    UpdateOrderStatusDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - pending
            - confirmed
            - preparing
            - out_for_delivery
            - delivered
            - cancelled
        note:
          type: string
      required:
        - status
...